requirements.txt
openai
pandas
matplotlib
numpy
python-dotenv
termcolor
sqlalchemy
requests
beautifulsoup4
replicate
fal-client
jsonschema
streamlit
selenium
lxml
sounddevice
soundfile
seaborn
pynput
pydub
playsound
simpleaudio
elevenlabs
pyperclip
llama-index-core
llama-index-llms-openai
llama-index-readers-file
llama-index-embeddings-openai
pygame
anthropic
e2b-code-interpreter
azure-storage-blob